@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO.Abstractions;
 using System.Reflection;
@@ -6,12 +6,13 @@ using BlackSlope.Api.Common.Configuration;
 using BlackSlope.Api.Common.Extensions;
 using BlackSlope.Api.Common.Middleware.Correlation;
 using BlackSlope.Api.Common.Middleware.ExceptionHandling;
-using BlackSlope.Api.Common.Version.Interfaces;
-using BlackSlope.Api.Common.Version.Services;
+using BlackSlope.Api.Common.Versioning.Interfaces;
+using BlackSlope.Api.Common.Versioning.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace BlackSlope.Api
 {
@@ -46,8 +47,7 @@ namespace BlackSlope.Api
             services.AddMovieValidators();
         }
 
-
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -63,11 +63,31 @@ namespace BlackSlope.Api
             app.UseHttpsRedirection();
 
             app.UseSwagger(HostConfig.Swagger);
+
+            app.UseRouting();
             app.UseCors("AllowSpecificOrigin");
+
             app.UseAuthentication();
+
             app.UseMiddleware<CorrelationIdMiddleware>();
             app.UseMiddleware<ExceptionHandlingMiddleware>();
-            app.UseMvc();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
+        }
+
+        private static void CorsConfiguration(IServiceCollection services)
+        {
+            services.AddCors(options =>
+            {
+                options.AddPolicy(
+                    "AllowSpecificOrigin",
+                    builder => builder.AllowAnyOrigin() // TODO: Replace with FE Service Host as appropriate to constrain clients
+                        .AllowAnyHeader()
+                        .WithMethods("PUT", "POST", "OPTIONS", "GET", "DELETE"));
+            });
         }
 
         // make a list of projects in the solution which must be scanned for mapper profiles
@@ -81,18 +101,6 @@ namespace BlackSlope.Api
 
             var serviceProvider = services.BuildServiceProvider();
             HostConfig = serviceProvider.GetService<HostConfig>();
-        }
-
-        private void CorsConfiguration(IServiceCollection services)
-        {
-            services.AddCors(options =>
-            {
-                options.AddPolicy(
-                    "AllowSpecificOrigin",
-                    builder => builder.AllowAnyOrigin() // TODO: Replace with FE Service Host as appropriate to constrain clients
-                        .AllowAnyHeader()
-                        .WithMethods("PUT", "POST", "OPTIONS", "GET", "DELETE"));
-            });
         }
     }
 }
