@@ -8,7 +8,7 @@ namespace BlackSlope.Api.Common.Extensions
     public static class EnumerationExtension
     {
         public static string GetDescription(this Enum value) =>
-            value.GetType()
+            value?.GetType()
                  .GetMember(value.ToString())
                  .FirstOrDefault()?
                  .GetCustomAttribute<DescriptionAttribute>()?
