@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics.Contracts;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Primitives;
 
@@ -10,8 +11,9 @@ namespace BlackSlope.Api.Common.Middleware.Correlation
 
         public Guid? Read(HttpContext context)
         {
-            StringValues correlationId = context.Request.Headers[CorrelationIdHeaderKey];
-            if (Guid.TryParse(correlationId.ToString(), out Guid result))
+            Contract.Requires(context != null);
+            var correlationId = context.Request.Headers[CorrelationIdHeaderKey];
+            if (Guid.TryParse(correlationId.ToString(), out var result))
             {
                 return result;
             }
@@ -23,6 +25,7 @@ namespace BlackSlope.Api.Common.Middleware.Correlation
 
         public void Write(HttpContext context, Guid correlationId)
         {
+            Contract.Requires(context != null);
             context.Response.Headers[CorrelationIdHeaderKey] = correlationId.ToString();
         }
     }
