@@ -6,9 +6,9 @@ using FluentValidation;
 
 namespace BlackSlope.Api.Operations.Movies.Validators
 {
-    public class CreateMovieViewModelValidator : BlackslopeValidator<CreateMovieViewModel>
+    public class CreateMovieViewModelValidatorCollection : BlackslopeValidatorCollection<CreateMovieViewModel>
     {
-        public CreateMovieViewModelValidator(IMovieService movieService)
+        public CreateMovieViewModelValidatorCollection(IMovieService movieService)
         {
             RuleFor(x => x).MustAsync(async (x, cancellationtoken) =>
                     !await movieService.CheckIfMovieExistsAsync(x.Title, x.ReleaseDate))
